refactor: put InMemorySessionStore in data package
